Pull a tiles dependency out of player.h
The end goal is eliminating unnecessary #ifdefs in header files very
commonly included.  This commit is a quick fix: better long-run
solutions include refactoring player_save_info so that the doll
information simply isn't part of it, or changing things so that the doll
information is loaded in non-tiles builds.